+++ R/regTable.R
@@ -312,11 +312,18 @@ regTable <- function (
 
 #' @param x regTable object.
 #' @param decimalPlaces Integer. Table entries will be rounded to this number
-#' of digits after the decimal point.
+#' of digits after the decimal point. By default, it is the value of 
+#' \code{getOption("Bullock.print.regTable.dp")}, which is set to{\NB}2 when 
+#' the Bullock package is loaded. You can change the default by running 
+#' \code{options("Bullock.print.regTable.dp") <- N} at any time, where
+#' \code{N} is your preferred number of digits after the decimal place.
 #' @param ... Arguments passed to \code{print.table()}.
 #' @importFrom rlang .data
 #' @export 
-print.regTable <- function (x, decimalPlaces = 2L, ...) {
+print.regTable <- function (
+  x, 
+  decimalPlaces = getOption("Bullock.print.regTable.dp"), 
+  ...) {
 
   # We are right-aligning the data columns. Each column-tier has a 
   # "column-tier name," and we are right-aligning this name, too. In some